Rename cjwstate/modules/message_catalogs
